configure Automatic-Module-Name to be javalin.pac4j
Sadly, some dependencies are not correctly modularized. Because Javalin still depends on those, one cannot simply create a module-info as that would not reliably pick up those dependencies across build tools/IDEs. By making this library an automatic module we can circumvent this shortcoming for the moment by compiling without module support but naming this library as a module for users