revert pac4j version downgrade
By upgrading pac4j to a version that splits out pac4j-jee we face the problem of having two modules (pac4j.core and pac4j.jee) that both declare some packages (eg org.pac4j.core.http.adapter). That creates effectively a split package which does not work with java modules. To work around this, we remove dependency on jee completely by implementing the core interfaces in face of javalin by mimicing the way jee works.
While this works and compiles, it leaves us with a necessasity to depend on javax.servlet.api (because javalin builds on it).
This compiles with maven, but users may find it difficult to depend on javax.servlet.api themselves.

Apart from it, the amount of glueing code seem to outweigh the benefits of having a module-info plus no jee anymore.
Thus it seems to be more usable to just include an Automatic-Module-Name in the manifest and drop the attempt to build a module-info.